requirements.txt
prettytable==3.4.1
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
